backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29
pydantic==2.6.4
pydantic[email]==2.6.4

passlib[bcrypt]==1.7.4
bcrypt==4.1.2

python-multipart==0.0.9

python-jose==3.3.0

python-dotenv==1.0.0
httpx==0.25.2
google-auth==2.20.0
itsdangerous==2.1.2
requests==2.31.0
